part2_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
httpx==0.25.2
uagents==0.1.0
pillow==10.1.0
pydub==0.25.1
requests==2.31.0
